Remove .float() from reset() too in VecPyTorch

+++ a2c_ppo_acktr/envs.py
@@ -167,7 +167,7 @@ class VecPyTorch(VecEnvWrapper):
 
     def reset(self):
         obs = self.venv.reset()
-        obs = torch.from_numpy(obs).float().to(self.device)
+        obs = torch.from_numpy(obs).to(self.device)
         return obs
 
     def step_async(self, actions):
